% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_velocity.R
\name{calc_velocity}
\alias{calc_velocity}
\title{Compute the velocity of K and I}
\usage{
calc_velocity(indicators, M, Begining_Time)
}
\arguments{
\item{indicators}{dataframe that stores indicators calculate by get_indicators():
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item confirmed: the daily confirmed cases at the given date
\item recovered: the daily recovered cases at the given date
\item deaths: the daily deaths at the given daten
\item cum_confirmed: the cumulative confirmed cases
\item inhospitals: The number of infectious cases in hospital
\item infection_rate: The daily infection rate
\item removed_rate: the daily removed rate
}}

\item{M}{the selection of time window.}

\item{Begining_Time}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
A list contains 4 elements:
\itemize{
\item infection_rate_velocity: the velocity of K.
\item removed_rate_velocity: the velocity of I.
\item corr.infection: vvector of corrected M_K and T_K.
\item corr.removed: vector of corrected M_I and T_I.
These 2 corrected vector in the format of "c(T, M)" used to calculate
the above 2 indicators which may be needed in later calculation.
}
}
\description{
Compute the velocity of infection rate K change and completion rate I change.
}
\examples{
Begining_Time <- "2020-01-29"
M <- 5
indicators <- get_indicators(DemoPreTurningPointsCOVID19::COVID19_CN)
velocity <- calc_velocity(indicators, M, Begining_Time)
}
