% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVelocity.R
\name{CalculateVelocity}
\alias{CalculateVelocity}
\title{Compute the velocity of K and I}
\usage{
CalculateVelocity(
  date,
  confirmed,
  inhospitals,
  infectionrate,
  removedrate,
  M,
  ST
)
}
\arguments{
\item{date}{the exact day in the formate "\%y-\%m-\%d" as a character}

\item{confirmed}{the daily confirmed cases at the given date}

\item{inhospitals}{The number of infectious cases in hospital}

\item{infectionrate}{The daily infection rate}

\item{removedrate}{the daily removed rate}

\item{M}{the selection of time window.}

\item{ST}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
A list contains InfectionRateVelocity, RemovedRateVelocity,
and the final T.removed and M.removed used to calculate RemovedRateVelocity,
which may be needed in later calculation.
}
\description{
Compute the velocity of infection rate K change and completion rate I change.
}
