% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_removed_rate.R
\name{corr_removed_rate}
\alias{corr_removed_rate}
\title{Correct removed rate}
\usage{
corr_removed_rate(velocity, indicators, pre_data, M, Begining_Time)
}
\arguments{
\item{velocity}{A list contains 4 elements:
\itemize{
\item infection_rate_velocity: the velocity of K.
\item removed_rate_velocity: the velocity of I.
\item corr.infection: vvector of corrected M_K and T_K.
\item corr.removed: vector of corrected M_I and T_I.
}}

\item{indicators}{dataframe with variables:
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item confirmed: the daily confirmed cases at the given date
\item recovered: the daily recovered cases at the given date
\item deaths: the daily deaths at the given daten
\item cum_confirmed: the cumulative confirmed cases
\item inhospitals: The number of infectious cases in hospital
\item infection_rate: The daily infection rate
\item removed_rate: the daily removed rate
}}

\item{pre_data}{dataframe with 4 predicted indicators:
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item infection_rate.pre: prediction result of infection rate
\item removed_rate.pre: prediction result of removed_rate
\item inhospitals.pre: prediction result of inhospitals
\item confirmed.pre: prediction result of confirmed cases
All of these data is calculated from the Begining_Time.
}}

\item{M}{the selection of time window.}

\item{Begining_Time}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
A list contains 4 elements which go through the correction:
\itemize{
\item infection_rate_velocity: the velocity of K.
\item removed_rate_velocity: the velocity of I.
\item corr.infection: vvector of corrected M_K and T_K.
\item corr.removed: vector of corrected M_I and T_I.
These 2 corrected vector in the format of "c(T, M)" used to calculate
the above 2 indicators which may be needed in later calculation.
}
}
\description{
if future removed_rate is larger than 1, than correct T and M
to make removed_rate less than 1.
}
\examples{
Begining_Time <- "2020-01-29"
M <- 5
indicators <- get_indicators(DemoPreTurningPointsCOVID19::COVID19_CN)
velocity <- calc_velocity(indicators, M, Begining_Time)
pre_data <- get_future_indicators(indicators, velocity, Begining_Time)
new_velocity <- corr_removed_rate(velocity, indicators, pre_data, M, Begining_Time)
}
