% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COVID19_CN}
\alias{COVID19_CN}
\title{COVID-19 Data of China Mainland beyond Hubei}
\format{
A data frame with 39 rows and 5 variables:
\describe{
\item{date}{the day t}
\item{cumulated_confirmed}{the cumulative confirmed cases up to the given day t}
\item{daily_confirmed}{the daily confirmed cases at day t}
\item{daily_recoverd}{the daily recovered ones at day t}
\item{daily_death}{the daily deaths at day t}
}
}
\source{
\url{https://github.com/Vicky-Zh/Tracking_and_forecasting_milepost_moments_of_COVID-19/blob/master/data\%20and\%20code\%20for\%20paper/Data\%20of\%20China\%20Mainland\%20Beyond\%20Hubei.csv}
}
\usage{
COVID19_CN
}
\description{
A dataset containing the date, cumulated_confirmed and other attributes
of 39 days from 2020-1-22 to 2020-2-29.
}
\keyword{datasets}
