% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{Compute the iconic indicators}
\usage{
get_indicators(RawData)
}
\arguments{
\item{RawData}{dataframe with data and four ordered variables, i.e.
\itemize{
\item the day t,
\item the cumulative confirmed cases up to the given day t,
\item the daily confirmed cases at day t,
\item the daily recovered ones at day t,
\item the daily deaths at day t.
}}
}
\value{
dataframe with variables:
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item confirmed: the daily confirmed cases at the given date
\item recovered: the daily recovered cases at the given date
\item deaths: the daily deaths at the given daten
\item cum_confirmed: the cumulative confirmed cases
\item inhospitals: The number of infectious cases in hospital
\item infection_rate: The daily infection rate
\item removed_rate: the daily removed rate
}
}
\description{
Compute the iconic indicators and return as dataframe
}
\examples{
indicators <- get_indicators(DemoPreTurningPointsCOVID19::COVID19_CN)
}
