% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_milepost.R
\name{get_milepost}
\alias{get_milepost}
\title{Get Milepost}
\usage{
get_milepost(pre_data, Begining_Time)
}
\arguments{
\item{pre_data}{dataframe with 4 predicted indicators:
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item infection_rate.pre: prediction result of infection rate
\item removed_rate.pre: prediction result of removed_rate
\item inhospitals.pre: prediction result of inhospitals
\item confirmed.pre: prediction result of confirmed cases
All of these data is calculated from the Begining_Time.
And it is always generated by function prediction().
}}

\item{Begining_Time}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
dataframe contains 4 milepost T.1, T.2, Z.1, Z.2,
which must be in the formate "\%y-\%m-\%d" as a character.
}
\description{
Get 4 milepost moments T.1, T.2, Z.1 and Z.2
}
\examples{
Begining_Time <- "2020-01-29"
M <- 5
indicators <- get_indicators(DemoPreTurningPointsCOVID19::COVID19_CN)
velocity <- calc_velocity(indicators, M, Begining_Time)
pre_data <- get_future_indicators(indicators, velocity, Begining_Time)
mileposts <- get_milepost(pre_data, Begining_Time)
}
