% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Iconicfun.R
\name{Iconicfun}
\alias{Iconicfun}
\title{Compute the iconic indicators}
\usage{
Iconicfun(wd)
}
\arguments{
\item{wd}{dataframe with data and four variables, i.e.
the cumulative confirmed cases up to the given day t,
the daily confirmed cases at day t,
the daily recovered ones and the daily deaths at day t，}
}
\value{
dataframe with variables
date: the exact day in the formate "\%y-\%m-\%d" as a character
confirmed: the daily confirmed cases at the given date
recovered: the daily recovered cases at the given date
deaths: the daily deaths at the given date
cumconfirmed: the cumulative confirmed cases
inhospitals: The number of infectious cases in hospital
infectionrate: The daily infection rate
removedrate: the daily removed rate
}
\description{
Compute the iconic indicators
}
