% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_predict.R
\name{period_predict}
\alias{period_predict}
\title{Prediction over a period of time}
\usage{
period_predict(datafile, M, Begining_Time, period)
}
\arguments{
\item{datafile}{it can be the path of data file as Charactors or
be the raw data read from .csv as a dataframe.}

\item{M}{the selection of time window.}

\item{Begining_Time}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}

\item{period}{the number of days in the selected time period.}
}
\value{
List contains as many sublist as the value of param period.
Each sublist is the result of a prediciton with the given time t as
begining time. And each sublist is named with the begining time of predicion.
}
\description{
Take each day of a period of time as the starting point of prediction
and get the result.
}
\examples{
Begining_Time <- "2020-01-29"
M <- 5
period <- 32
period_predict(DemoPreTurningPointsCOVID19::COVID19_CN, M,
Begining_Time, period)
}
