% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeriodForcast.R
\name{PeriodForcast}
\alias{PeriodForcast}
\title{Forcast over a period of time}
\usage{
PeriodForcast(filepath, M, ST, period)
}
\arguments{
\item{filepath}{path of the target data file.}

\item{M}{the selection of time window.}

\item{ST}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}

\item{period}{the number of days in the selected time period.}
}
\value{
A dataframe contains begining time of prediction,
and T.2, Z.1, Z.2 in the formate "\%y-\%m-\%d" as a character.
BeginingTime: the begining time of prediction.
T.2: the second turning point
Z.1: the first 'zero' point
Z.2: the second 'zero' point
}
\description{
Forcast over a period of time. Each day in the period is taken as
the starting time of prediction.
}
