% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Prediction}
\alias{Prediction}
\title{Predict future indicators and get 3 turning points}
\usage{
Prediction(
  date,
  confirmed,
  inhospitals,
  infectionrate,
  removedrate,
  InfectionRateVelocity,
  RemovedRateVelocity,
  ST
)
}
\arguments{
\item{date}{the exact day in the formate "\%y-\%m-\%d" as a character.}

\item{confirmed}{the daily confirmed cases at the given date.}

\item{inhospitals}{The number of infectious cases in hospital.}

\item{infectionrate}{The daily infection rate.}

\item{removedrate}{the daily removed rate.}

\item{InfectionRateVelocity}{the velocity of infection rate change.}

\item{RemovedRateVelocity}{the velocity of complerion rate change.}

\item{ST}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
A dataframe contains prediction result of removedrate, inhospitals,
and T.2, Z.1, Z.2.
}
\description{
Predict future infectionrate, removedrate, E_t, inhospitals
and get T.2, Z.1 and Z.2 at the same time.
}
