% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalPrediction.R
\name{totalPrediction}
\alias{totalPrediction}
\title{Integrate functions and handle a special situation (removedrate > 1)}
\usage{
totalPrediction(wd, M, ST)
}
\arguments{
\item{wd}{dataframe with data and four variables, i.e.
the cumulative confirmed cases up to the given day t,
the daily confirmed cases at day t,
the daily recovered ones and the daily deaths at day t}

\item{M}{the selection of time window.}

\item{ST}{the selection of begining time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
A dataframe contains prediction result of removedrate, inhospitals,
and T.2, Z.1, Z.2.
}
\description{
Integrate functions above, and handle a special situation (removedrate > 1).
}
