% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_future_indicators.R
\name{get_future_indicators}
\alias{get_future_indicators}
\title{Get future indicators}
\usage{
get_future_indicators(indicators, velocity, Beginning_Time)
}
\arguments{
\item{indicators}{dataframe that stores indicators calculate by get_indicators():
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item confirmed: the daily confirmed cases at the given date
\item recovered: the daily recovered cases at the given date
\item deaths: the daily deaths at the given daten
\item cum_confirmed: the cumulative confirmed cases
\item active_cases: The number of infectious cases in hospital
\item infection_rate: The daily infection rate
\item removed_rate: the daily removed rate
}}

\item{velocity}{A list contains infection_rate_velocity, removed_rate_velocity,
and 2 corrected vector c(T, M) used to calculate the above 2 indicators
which may be needed in later calculation.}

\item{Beginning_Time}{the selection of beginning time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
dataframe with 4 predicted indicators:
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item infection_rate.pre: prediction result of infection rate
\item removed_rate.pre: prediction result of removed_rate
\item active_cases.pre: prediction result of active_cases
\item confirmed.pre: prediction result of confirmed cases
All of these data is calculated from the Beginning_Time.
}
}
\description{
Get future infection_rate, removed_rate, active_cases and confirmed cases.
}
\examples{
Beginning_Time <- "2020-01-29"
M <- 5
indicators <- get_indicators(DemoPreTurningPointsCOVID19::COVID19_CN)
velocity <- calc_velocity(indicators, M, Beginning_Time)
pre_data <- get_future_indicators(indicators, velocity, Beginning_Time)
}
