% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\title{Prediciton}
\usage{
prediction(RawData, M, Beginning_Time)
}
\arguments{
\item{RawData}{dataframe with data and four ordered variables, i.e.
\itemize{
\item the day t,
\item the cumulative confirmed cases up to the given day t,
\item the daily confirmed cases at day t,
\item the daily recovered ones at day t,
\item the daily deaths at day t.
}}

\item{M}{the selection of time window.}

\item{Beginning_Time}{the selection of beginning time,
which must be in the formate "\%y-\%m-\%d" as a character.}
}
\value{
List contains important info. in the calculation:
\itemize{
\item Beginning_Time: Hyper Param. the selection of beginning time,
which must be in the formate "\%y-\%m-\%d" as a character.
\item M: Hyper Param. the selection of time window.
\item indicators: the iconic indicators generated from real data.
\item velocity: A list contains 4 elements:
\itemize{
\item infection_rate_velocity: the velocity of K.
\item removed_rate_velocity: the velocity of I.
\item corr.infection: vvector of corrected M_K and T_K.
\item corr.removed: vector of corrected M_I and T_I.
}
\item pre_indicators: dataframe with 4 future indicators:
\itemize{
\item date: the exact day in the formate "\%y-\%m-\%d" as a character
\item infection_rate.pre: prediction result of infection rate
\item removed_rate.pre: prediction result of removed_rate
\item active_cases.pre: prediction result of active_cases
\item confirmed.pre: prediction result of confirmed cases
}
\item mileposts: dataframe contains 4 milepost T.1, T.2, Z.1, Z.2,
which must be in the formate "\%y-\%m-\%d" as a character.
}
}
\description{
Integrate functions above, and handle a special situation (removedrate > 1).
}
\examples{
Beginning_Time = "2020-01-29"
M <- 5
result <- prediction(COVID19_CN, M, Beginning_Time)
}
